Implicitly cast int multiplication to long

+++ src/main/java/net/aoba/misc/ModuleUtils.java
@@ -68,7 +68,7 @@ public class ModuleUtils {
 
                     return new ChunkPos(x, z);
 
-                }).limit(diameter * diameter)
+                }).limit((long) diameter * diameter)
                 .filter(c -> AobaClient.MC.world.isChunkLoaded(c.x, c.z))
                 .map(c -> AobaClient.MC.world.getChunk(c.x, c.z)).filter(Objects::nonNull);
